Test for enabled org NPI
